requirements.txt
os
Flask
torch
pillow
numpy
Python-IO
opencv-python
ultralytics